Fix crash
