+++ app/src/main/java/org/meowcat/edxposed/manager/SettingsFragment.java
@@ -30,9 +30,9 @@ import java.io.IOException;
 import java.util.Objects;
 
 import static org.meowcat.edxposed.manager.SettingsActivity.getDarkenFactor;
-import static org.meowcat.edxposed.manager.StatusInstallerFragment.getArch;
 import static org.meowcat.edxposed.manager.XposedApp.WRITE_EXTERNAL_PERMISSION;
 import static org.meowcat.edxposed.manager.XposedApp.darkenColor;
+import static org.meowcat.edxposed.manager.XposedApp.getArch;
 import static org.meowcat.edxposed.manager.XposedApp.getPreferences;
 import static org.meowcat.edxposed.manager.adapter.LogsHelper.isMainUser;
