+++ app/src/main/java/org/meowcat/edxposed/manager/XposedApp.java
@@ -45,8 +45,11 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Objects;
 
+import dalvik.system.VMRuntime;
 import de.robv.android.xposed.installer.util.InstallZipUtil;
 
+import static android.os.Build.SUPPORTED_32_BIT_ABIS;
+import static android.os.Build.SUPPORTED_64_BIT_ABIS;
 import static org.meowcat.edxposed.manager.MeowCatApplication.TAG;
 import static org.meowcat.edxposed.manager.adapter.AppHelper.FORCE_WHITE_LIST_MODULE;
 
@@ -63,6 +66,26 @@ public class XposedApp extends de.robv.android.xposed.installer.XposedApp implem
     private boolean mIsUiLoaded = false;
     private SharedPreferences mPref;
     private Activity mCurrentActivity = null;
+    public static String CPU_ABI;
+    private static String CPU_ABI2;
+
+    public static String getArch() {
+        if (CPU_ABI.equals("arm64-v8a")) {
+            return "arm64";
+        } else if (CPU_ABI.equals("x86_64")) {
+            return "x86_64";
+        } else if (CPU_ABI.equals("mips64")) {
+            return "mips64";
+        } else if (CPU_ABI.startsWith("x86") || CPU_ABI2.startsWith("x86")) {
+            return "x86";
+        } else if (CPU_ABI.startsWith("mips")) {
+            return "mips";
+        } else if (CPU_ABI.startsWith("armeabi-v5") || CPU_ABI.startsWith("armeabi-v6")) {
+            return "armv5";
+        } else {
+            return "arm";
+        }
+    }
 
     public static XposedApp getInstance() {
         return mInstance;
@@ -185,6 +208,19 @@ public class XposedApp extends de.robv.android.xposed.installer.XposedApp implem
 
     public void onCreate() {
         super.onCreate();
+
+        final String[] abiList;
+        if (VMRuntime.getRuntime().is64Bit()) {
+            abiList = SUPPORTED_64_BIT_ABIS;
+        } else {
+            abiList = SUPPORTED_32_BIT_ABIS;
+        }
+        CPU_ABI = abiList[0];
+        if (abiList.length > 1) {
+            CPU_ABI2 = abiList[1];
+        } else {
+            CPU_ABI2 = "";
+        }
 
         final ApplicationInfo appInfo = getApplicationInfo();
         BASE_DIR = appInfo.deviceProtectedDataDir + "/";
